added skeleton for testing Solver and Extractor
